refactor(api): expose a JavaScript-free HTTP tool client
